requirements.txt
openai>=1.40.0
rich>=13.7.0
click>=8.1.7
python-dotenv>=1.0.0
